refactor: 테스트 class 구조 변경
- GameTest에서 CarsTest, GameTest, ParserTest로 나눔
